Special Traits (#646)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

A lot of these made by Steiner, big thanks to him!

Use "Be Special" in your prefs to roll for a special trait, and when you
spawn, you'll Be Special, which could be something negative, positive or
neutral

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Adds more randomization and spices up the rounds with interesting
changes, that add RP hooks straight from the start of the round, it is
also fully opt-in

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

---------

Co-authored-by: Götz <14265868+Wildwestmars@users.noreply.github.com>