21 New Specials (#712)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

21 new Specials have been added.
What could they be? You can find out in-game... or spoil the experience
for yourself and read the code here

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->