Primal Savagery (Poison Bite), Poison Spray, Ray of Frost
Added:
Poison Spray: takes a container you are holding and sprays it's contents in a cloud. Consumes the reagents in the container.

Primal Savagery: adds a buff for 30 seconds which causes your bites to inject venom and soporific into you target. added a new trait POISONBITE and made sure it works on all mobs.

Ray of Frost: a slower projectile that slows for 3 seconds on hit.

More scrolls

Changes:
More cooldown changes